delete extra export button and moved to nav bar
