+++ package.json
@@ -15,7 +15,7 @@
     "start-dev": "cross-env NODE_ENV=development webpack serve --hot --host 0.0.0.0 --config=./webpack.dev.config.js --mode development"
   },
   "keywords": [],
-  "author": "",
+  "author": "Rain Hsu, Bryanna DeJesus, Jake Pino and Ronak Hirpara",
   "license": "MIT",
   "devDependencies": {
     "@babel/core": "^7.16.7",
@@ -49,6 +49,7 @@
     "react-dom": "^17.0.2",
     "react-modal": "^3.14.4",
     "react-spinners": "^0.11.0",
+    "react-tabs": "^3.2.3",
     "style-loader": "^3.3.1",
     "tailwindcss": "^3.0.15",
     "ts-loader": "^9.2.6",
