+++ src/components/DnD.jsx
@@ -8,7 +8,7 @@ import CustomComponents from './CustomComponents';
 const DnD = () => {
   return (
     <div className='dndContainer'>
-        <p id='app-name'>App Name</p>
+        <p id='app-name'>fflow.</p>
         <CompCreator />
         <TagCreator />
         <DragList />
