+++ src/components/DragList.jsx
@@ -3,7 +3,6 @@ import '../stylesheets/DragList.css';
 import { Droppable, Draggable } from 'react-beautiful-dnd';
 import { useDispatch } from 'react-redux';
 import { useSelector } from 'react-redux';
-import drag from '../img/drag.png';
 
 const DragList = () => {
   const items = useSelector(state => state.tags.tagList);
