+++ src/index.html
@@ -7,7 +7,7 @@
   <link rel="stylesheet" href="https://unpkg.com/@themesberg/flowbite@1.3.0/dist/flowbite.min.css" />
   <link rel="stylesheet" href="https://rsms.me/inter/inter.css">
   <script src="https://kit.fontawesome.com/4a4720e8ee.js" crossorigin="anonymous"></script>
-  <title>Our OSP</title>
+  <title>fflow</title>
 </head>
 <body>
   <div id='root'></div>
