+++ src/redux/tagsSlice.js
@@ -15,6 +15,8 @@ const initialState = {
     h1Tag: 'Header 1',
     h2Tag: 'Header 2',
     h3Tag: 'Header 3',
+    tableTag: 'Table',
+    tHeadTag: 'THead'
   },
 };
